requirements.txt
# ============================================================
# PYTHON DEPENDENCIES
# Unilever Data Warehouse Pipeline
# ============================================================

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database Connectivity
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data Generation
faker>=19.0.0

# ETL & Pipeline
airflow>=2.7.0

# Monitoring & Alerting
prometheus-client>=0.17.0
grafana-api>=1.0.0

# Logging
python-json-logger>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
